requirement.txt
asgiref==3.2.7
bson==0.5.8
certifi==2020.4.5.1
chardet==3.0.4
dataclasses==0.7
defusedxml==0.6.0
Django==2.2.12
django-allauth==0.41.0
django-extensions==2.2.9
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djongo==1.3.2
idna==2.9
oauthlib==3.1.0
Pillow==7.1.1
PyJWT==1.7.1
pymongo==3.10.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.2.4
urllib3==1.25.8
